Use a powershell script to manage focus on Windows
Fixes #4

Sublime Text's `bring_to_front` politely asks for the focus which is
often not enough when a browser has input focus.

But a) if it does it is instant and b) it helps when all windows are
minimized.

As we cannot check if `bring_to_front` actually did what we want, use
the scripts we always had here too, unconditionally.

For Windows, replace the used method with a powershell script which
works on newer Windows versions without installing the rather
idiosyncratic "nircmd".

Unfortunately Linux still needs a tool ("xdotool").